Move PaperAPI handlers to a new route
